cleanup of strategy
